Ensure that the Jenkins build timestamp is interpreted as UTC
Recently, either the Jenkins build timestamp or Chrome's Date.parse
implementation changed so that the timestamp is now incorrectly
interpreted as local time rather than UTC.  Firefox does the same.  Safari
still reads it as UTC.  The string is reporting the UTC time, but contains
no time zone information, so that its interpretation is ambiguous and
implementation-dependent in practice.

This appends a 'Z' to the date string so that it's interpreted correctly
across browser JS implementations.

Change-Id: I1955fde80906f9d5a1eaf50136020a1dfef1b787